+++ gepy/functions/truncations.py
@@ -1,0 +1,14 @@
+from gepy.decorators import symbol
+import math
+
+@symbol("floor")
+def floor(i):
+    return math.floor(i)
+
+@symbol("ceil")
+def ceil(i):
+    return math.ceil(i)
+
+@symbol("round")
+def op_round(i):
+    return round(i)