+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog for gepy
 ## Version 0.1
+- Implementation of rounding and comparison functions
 - Implementation of basic arithmetic functions
 - Implementation of basic multigene chromosomes
 - Implementation of more important evolution operators: mutation and head inversion